refactor(app): move some old pipette types
These look pretty silly sitting next to each other. We should probably
fix that
