+++ app/src/redux/pipettes/__fixtures__/index.ts
@@ -2,18 +2,16 @@
 import { fixtureP10Single } from '@opentrons/shared-data/pipette/fixtures/name'
 import type {
   AttachedPipette,
+  PipetteInfo,
   PipetteSettings,
   PipetteSettingsFieldsMap,
+  PipetteInformation,
 } from '../types'
 import type {
   RobotApiResponse,
   RobotApiResponseMeta,
 } from '../../robot-api/types'
 import { mockTipRackDefinition } from '../../custom-labware/__fixtures__'
-import type {
-  PipetteInfo,
-  PipetteInformation,
-} from '/app/organisms/Devices/hooks'
 import type { PipetteData } from '@opentrons/api-client'
 
 export const mockRobot = { name: 'robot', ip: '127.0.0.1', port: 31950 }
