+++ app/src/redux/pipettes/index.ts
@@ -7,3 +7,4 @@
 export * from './actions'
 export * from './constants'
 export * from './selectors'
+export type * from './types'
